+++ greenfield/eldenring/tables/missable_locations.py
@@ -299,4 +299,4 @@ MISSABLE_LOCATIONS = {
     7900274: 'questline',
 }
 
-_GEN_STAMP = {'inputs_hash': 'sha256:023b73ed947ed482b23511c8a60d2c1816503ba5ee09daa01ad52d394f67dcbd', 'module': 'missable_locations.py', 'body_sha256': 'sha256:73459ba6761edd7bcedad6afcadd9e31e11089061beeea6813cb6519b00eeb85'}
+_GEN_STAMP = {'inputs_hash': 'sha256:0e6f9cc5d4d7d9988bc47ce3f5a0e912983b53ab5633cb85b75f9c5db36a2bcc', 'module': 'missable_locations.py', 'body_sha256': 'sha256:73459ba6761edd7bcedad6afcadd9e31e11089061beeea6813cb6519b00eeb85'}